# Add `unique!`