--- a/base/exports.jl
+++ b/base/exports.jl
@@ -716,6 +716,7 @@ export
     symdiff,
     union!,
     union,
+    unique!,
     unique,
     values,
     valtype,