--- a/base/exports.jl
+++ b/base/exports.jl
@@ -720,6 +720,7 @@ export
     symdiff,
     union!,
     union,
+    unique!,
     unique,
     values,
     valtype,