--- a/base/set.jl
+++ b/base/set.jl
@@ -215,6 +215,7 @@ julia> A
 function unique!(A::AbstractVector)
     isempty(A) && return A
 
+    sorted = false
     try
         sorted = issorted(A) || issorted(A, rev=true)
     catch
@@ -227,7 +228,7 @@ function unique!(A::AbstractVector)
         idxs = eachindex(A)
         y = first(A)
         state = start(idxs)
-        _, state = next(idxs, state)
+        j, state = next(idxs, state)
         for i in idxs
             x = A[i]
             if !isequal(x, y)
@@ -241,7 +242,7 @@ function unique!(A::AbstractVector)
         # we have seen so far.
         seen = Set{eltype(A)}()
         idxs = eachindex(A)
-        state = start(idxs)
+        i = state = start(idxs)
         for x in A
             if x ∉ seen
                 push!(seen, x)