… NEWS.md
… doc/src/stdlib/collections.md
… test/TestHelpers.jl
… test/sets.jl