Added a `unique!` function that creates a fast path for sorted data.
Also added some tests for the function.

Continuation of #20576. Fixes https://github.com/JuliaLang/julia/issues/20549.