@JackDevine 2017-06-07
@StefanKarpinski @mbauman , I am not sure why travis didn't pass. Would it be easier if I just deleted this branch and started again?