@nalimilan 2017-06-07
No, keeping this PR is fine. I've restarted the build (it had failed due to a timeout).