@nalimilan 2017-06-07
Hmm, looks like another unrelated error:
```
From worker 1:		From worker 1:	Worker 5 failed running test spawn:

Some tests did not pass: 103 passed, 0 failed, 1 errored, 4 broken.spawn: Error During Test

  Got an exception of type LoadError outside of a @test

  LoadError: InterruptException:

  while loading /private/tmp/julia/share/julia/test/spawn.jl, in expression starting on line 425

```
https://gist.github.com/nalimilan/6783ae43c596828975ebd0c26f5e9b73

I've restarted the build again.