@JackDevine 2017-06-07
Thanks for all of your help people, I have rerun the benchmarks on my windows machine, which has the following specs:
64 bit OS,
Intel i7-4790K CPU @4.00GHz,
16.0GB ram

```julia
using BenchmarkTools

srand(42)
A = [rand(1:10) for _ in 1:100_000]
srand(42)
B = [rand(1:1000) for _ in 1:100_000]

println("Test performance on a large vector with not many unique elements.")
println("Non mutating unique():")
@btime unique(A)
println("Mutating unique()\n")
@btime unique!(A)

println("Test performance on a large vector with quite a few unique elements.")
println("Non mutating unique():")
@btime unique(B)
println("Mutating unique():")
@btime unique!(B)

Test performance on a large vector with not many unique elements.
Non mutating unique():
  1.046 ms (14 allocations: 1.19 KiB)
Mutating unique()
  1.873 μs (11 allocations: 960 bytes)
Test performance on a large vector with quite a few unique elements.
Non mutating unique():
  1.063 ms (59 allocations: 67.95 KiB)
Mutating unique():
  48.239 μs (50 allocations: 51.59 KiB)
```
So we have a factor 1000 speed up as well as a reduction in allocations.