@JackDevine 2017-06-08
Also, if I test the performance of the `@inbounds`, then the best I can get is the following:

```julia
srand(42)
B = [rand() for _ in 1:1_000_000]
@btime unique!(B)
```
With `@inbounds` I get
```
107.309 ms (122 allocations: 34.51 MiB)
```
and without it I get
```
115.333 ms (122 allocations: 34.51 MiB)
```
Which is not a great improvement, things don't get much better if I use `@inbounds begin ... end` to wrap both of the loops. Maybe we should consider not using `@inbounds`?