@nalimilan 2017-06-08
> Is it complaining that I pushed too many commits too quickly?

That's not an issue, it just means that a build was skipped because a new commit was pushed before it started.

Regarding `@inbounds`, better keep it even if the difference is small, it doesn't hurt. Also, you can put it right before `for`, which will enable it in more places than currently (in both branches).