@tkelman 2017-06-08
> Regarding `@inbounds`, better keep it even if the difference is small, it doesn't hurt

It does hurt if a user-defined type hasn't implemented its `eachindex` / `getindex` combination correctly and gets segfaults instead of a bounds error.