@nalimilan 2017-06-08
> It does hurt if a user-defined type hasn't implemented its eachindex / getindex combination correctly and gets segfaults instead of a bounds error.

Yeah, but an incorrectly implemented custom type will get bugs in many other places anyway. From https://github.com/JuliaLang/julia/issues/20469 I understood that the policy is that `AbstractArray` implementations which use the `@boundscheck` annotation are responsible for getting it right.