@JackDevine 2017-06-09
Just to prove that the OffsetArray stuff is working, if you have the code in TestHelpers and you include
```julia
Base.resize!(A::OffsetVector, nl::Integer) = resize!(A.parent, nl)
Base.done(A::OffsetVector, ind::Integer) = done(A.parent, ind - A.offsets[1])
```
Then the following tests work:
```julia
u = OffsetArray([1,2,2,3,5,5], -1)
unique!(u)
@assert u == OffsetArray([1,2,3,5], -1)

u = OffsetArray([5,5,4,4,2,2,0,-1,-1], -1)
unique!(u)
@assert u == OffsetArray([5,4,2,0,-1], -1)

u = OffsetArray([1,1,2,3,1,4,6], -1)
unique!(u)
@assert u == OffsetArray([1,2,3,4,6], -1)

u = OffsetArray([1,1,3,1,4,6], -1)
unique!(u)
@assert u == OffsetArray([1,3,4,6], -1)

u = OffsetArray(["w","we","w",5,"r",5,5], -1)
unique!(u)
@assert u == OffsetArray(["w","we",5,"r"], -1)
```

Something that I found when trying to make OffsetArrays work:
```julia
julia> unique([1,"w",1,3])
 3-element Array{Any,1}:
 1   
  "w"
 3

julia> unique!([1,"w",1,3])
 MethodError: no method matching isless(::String, ::Int64)
 Closest candidates are:
  isless(::AbstractFloat, ::Real) at operators.jl:98
  isless(::Real, ::Real) at operators.jl:266
  isless(::AbstractString, ::AbstractString) at strings/basic.jl:124

Stacktrace:
 [1] issorted(::Array{Any,1}, ::Base.Order.ForwardOrdering) at .\sort.jl:54
 [2] unique!(::Array{Any,1}) at .\In[3]:4
```
So checking that the array is sorted is causing problems. Although maybe this is not such a big deal because if you are using Any arrays then you probably aren't interested in the possible performance gains from `unique!`.