@nalimilan 2017-06-09
Great! Just add the tests and the functions here, if they are identical `git` will rebase automatically once the other PR has been merged.

Though it's problematic to throw an error when some entries cannot be compared. I guess the only solution is to wrap the `issorted` calls inside `try... catch`. (We could also play with `method_exists`, but for that wouldn't work for `Any`, and even if a function exists it could throw an error.)