@nalimilan 2017-06-09
> Thanks for the tips @nalimilan , at the moment the try ... catch block is in a little bit of an awkward place. I was thinking that maybe it should be after the isempty call, but I wanted your opinion first.

Looks OK to me, but maybe you should just do `isempty(A) && return A` at the top of the function. Also I would rename `Aissorted` to something shorter like `sorted` since there's no ambiguity.