@JackDevine 2017-06-09
With regards to `@inbounds`, I did some benchmarks with the latest version of `unique!`. One version didn't have `@inbounds` and the other had `@inbounds begin ... end` wrapping both loops.
```julia
using BenchmarkTools
srand(42)
A = [rand(1:100_000) for _ in 1:1_000_000] |> sort
srand(42)
B = [rand(1:100_000) for _ in 1:1_000_000]

println("Test the sorted branch.")
@btime unique!(A)

@assert A == unique(A)

println("Test the not sorted branch.")
@btime unique!(B)

@assert B == unique(B)
```
With `@inbounds`
```
Test the sorted branch.
  5.915 ms (298442 allocations: 4.55 MiB)
Test the not sorted branch.
  7.158 ms (198995 allocations: 6.04 MiB)
```
Without `@inbounds`
```
Test the sorted branch.
  6.426 ms (298442 allocations: 4.55 MiB)
Test the not sorted branch.
  7.324 ms (198995 allocations: 6.04 MiB)
```