@tkelman 2017-06-10
instead of a try-catch, this could maybe use a dispatch approach similar to https://github.com/JuliaLang/julia/pull/22010